requirements_pip.txt
# Tiaozhanbei2.0 项目依赖
# 管道追踪与法兰识别系统

# ============================ 核心依赖 ============================
# 计算机视觉
opencv-python>=4.8.0
numpy>=1.21.0

# RealSense 相机支持
pyrealsense2>=2.54.0

# 串口通信
pyserial>=3.5

# ============================ 数据处理 ============================
# 科学计算
scipy>=1.9.0

# 点云处理 (可选，如果需要高级点云处理)
# open3d>=0.17.0

# ============================ 系统工具 ============================
# 参数解析和配置
argparse  # 通常内置

# 日志和时间
# logging  # 内置
# datetime  # 内置

# ============================ 开发工具 ============================
# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0

# 代码质量
black>=23.0.0
flake8>=6.0.0

# ============================ 可选依赖 ============================
# 如果需要图形界面
# tkinter  # 通常内置

# 如果需要高级数学运算
# sympy>=1.11.0

# 如果需要深度学习推理
# torch>=2.0.0
# torchvision>=0.15.0

# 如果需要机器学习
# scikit-learn>=1.3.0

# ============================ 平台特定 ============================
# Windows 特定
# pywin32>=306  # 仅 Windows

# Linux 特定依赖会自动处理

# ============================ 版本说明 ============================
# Python 版本要求: >= 3.8
# 操作系统: Windows 10/11, Ubuntu 18.04+, macOS 10.15+
# 相机: Intel RealSense D400 系列
# 机器人: DJI RoboMaster C 板
